Major Major

+++ carla/evaluation/catalog/distance.py
@@ -26,7 +26,7 @@ def l0_distance(delta: np.ndarray) -> List[float]:
     num_feature_changes = np.sum(
         difference_mask,
         axis=1,
-        dtype=np.float,
+        dtype=np.float32,
     )
     distance = num_feature_changes.tolist()
     return distance
@@ -46,7 +46,7 @@ def l1_distance(delta: np.ndarray) -> List[float]:
     List[float]
     """
     absolute_difference = np.abs(delta)
-    distance = np.sum(absolute_difference, axis=1, dtype=np.float).tolist()
+    distance = np.sum(absolute_difference, axis=1, dtype=np.float32).tolist()
     return distance
 
 
@@ -64,7 +64,7 @@ def l2_distance(delta: np.ndarray) -> List[float]:
     List[float]
     """
     squared_difference = np.square(np.abs(delta))
-    distance = np.sum(squared_difference, axis=1, dtype=np.float).tolist()
+    distance = np.sum(squared_difference, axis=1, dtype=np.float32).tolist()
     return distance
